+ session ttl read with int value
git-svn-id: file:///Users/ephigenia/Sites/ephFrame/repository@182 4856e329-c559-0410-8ab9-924015c2118c

+++ 0.2/ephFrame/lib/component/Session.php
@@ -68,7 +68,7 @@ class Session extends Hash {
 		$this->start();
 		$this->data = &$_SESSION;
 		if (!$this->ttl) {
-			$this->ttl = ini_get('session.gc_maxlifetime');
+			$this->ttl = (int) ini_get('session.gc_maxlifetime');
 		} else {
 			ini_set('session.gc_maxlifetime', $this->ttl);
 		}
